Добавляет хук useInputs
