+++ project/public/css/main.css
@@ -17,7 +17,7 @@
 }
 
 .login__error-text {
-  top: -47px;
+  top: -30px;
   left: 5px;
 }
